requirements.txt
apache-libcloud>=2.2.1
attrs>=18.0
backports.functools-lru-cache>=1.0.2 ; python_version < '3'
boto3>=1.9
botocore>=1.12
colorama>=0.4.1
coloredlogs>=10.0
enum34>=0.9
funcsigs>=1.0
furl>=2.0.0
future>=0.16.0
futures>=3.0.5 ; python_version < '3'
google-cloud-storage>=1.13.2
humanfriendly>=2.1
jsonmodels>=2.2
jsonschema>=2.6.0
numpy>=1.10
opencv-python>=3.2.0.8
pathlib2>=2.3.0
psutil>=3.4.2
pyhocon>=0.3.38
python-dateutil>=2.6.1
PyYAML>=3.12
requests-file>=1.4.2
requests>=2.18.4
six>=1.11.0
tqdm>=4.19.5
urllib3>=1.22
watchdog>=0.8.0
pyjwt>=1.6.4
plotly>=3.9.0
